Also move all the editor tests to be organized together
