+++ apps/test/unit/lib/levelbuilder/level-editor/DropletPaletteSelectorTest.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {mount} from 'enzyme';
 import CodeMirror from 'codemirror';
-import {expect} from '../../../util/deprecatedChai';
+import {expect} from '../../../../util/deprecatedChai';
 import DropletPaletteSelector from '@cdo/apps/lib/levelbuilder/level-editor/DropletPaletteSelector';
 
 describe('DropletPaletteSelector', () => {
