+++ apps/test/unit/lib/levelbuilder/script-editor/LessonCardTest.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {shallow} from 'enzyme';
-import {expect} from '../../../util/reconfiguredChai';
+import {expect} from '../../../../util/reconfiguredChai';
 import sinon from 'sinon';
 
 import {UnconnectedLessonCard as LessonCard} from '@cdo/apps/lib/levelbuilder/script-editor/LessonCard';
