+++ apps/test/unit/lib/levelbuilder/script-editor/LessonGroupsTest.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {shallow} from 'enzyme';
-import {expect} from '../../../util/reconfiguredChai';
+import {expect} from '../../../../util/reconfiguredChai';
 import {UnconnectedLessonGroups as LessonGroups} from '@cdo/apps/lib/levelbuilder/script-editor/LessonGroups';
 
 const defaultProps = {
