+++ apps/test/unit/lib/levelbuilder/script-editor/NewLessonGroupInputTest.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {shallow} from 'enzyme';
-import {expect} from '../../../util/reconfiguredChai';
+import {expect} from '../../../../util/reconfiguredChai';
 
 import {UnconnectedNewLessonGroupInput as NewLessonGroupInput} from '@cdo/apps/lib/levelbuilder/script-editor/NewLessonGroupInput';
