+++ apps/test/unit/lib/levelbuilder/script-editor/ScriptAnnouncementsEditorTests.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {shallow} from 'enzyme';
-import {assert} from '../../../util/deprecatedChai';
+import {assert} from '../../../../util/deprecatedChai';
 import ScriptAnnouncementsEditor from '@cdo/apps/lib/levelbuilder/script-editor/ScriptAnnouncementsEditor';
 
 const sampleAnnouncement = {
